renamed image files to work around weird problem.
The files ending in _ad2.jpg and _ad3.jpg were not seen by Firefox nor Chrome on localhost via PhpStorm, nor on UH's server.  Other filenames are in the same pattern, so why those two were the only ones with this problem is a mystery.  Shortening the names didn't help, and changing _ad.jpg to _ad1.jpg gave it the same problem.  However, changing the last digit to a letter fixed it, and changing to -ad-[123].jpg also fixed it.